feat: ship docker release image and ts migrate entrypoint
